backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.27
pydantic==2.6.1
bcrypt==4.1.2
python-dotenv==1.0.1
pytz==2024.1 
gunicorn
opencc-python-reimplemented
